bench-python-project/pyproject.toml
[project]
name = "orchestrator-benchmark-tasks"
version = "0.1.0"
description = "I/O-bound benchmark tasks using DuckDB, MinIO and Postgres"
requires-python = ">=3.11"

   dependencies = [
    "boto3>=1.35.0,<2.0.0",
       "duckdb>=0.10.0",
       "psycopg2-binary>=2.9.9",
       "requests>=2.31.0"
   ]

[tool.uv]
# Ensures uv resolves wheels correctly inside container
python-preference = "only-system"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"